nlp_serving/pyproject.toml
[project]
name = "nlp-serving"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.9.*"
dependencies = [
    "fastapi[all]>=0.105.0",
    "loguru>=0.7.2",
    "multidict>=6.1.0",
    "pandas==2.1.3",
    "pydantic-settings>=2.1.0",
    "pytorch-lightning==2.1.2",
    "sqlmodel>=0.0.14",
    "torch==2.2.0",
    "transformers==4.35.2",
]
